Change block_count() to uint64_t
With larger block sizes, the malloc() argument would overflow, resulting
in an error.

+++ pattern.h
@@ -36,7 +36,7 @@ struct pattern {
 	const char *desc;
 
 	// Returns the size of the memory buffer in blocks.
-	int (*block_count)();
+	uint64_t (*block_count)();
 
 	// Returns the next thing to do.
 	struct cmd (*next_cmd)();
